Updated tic-tac-toe algorithm to include ML
included a basic ML learning algorithm to update its weights as it trains with the teacher data. Include auto generate teacher training data for the ML algorithm to use. Updated player manager to allow the user to play with the ML Model. notes: need to update the features of the algorithm for the model to play better.